Renamed publicHome folder to home
